Create action.yml
Step 3.  Add an action file.  I think the directory is the container for the GitHub action.  It contains the Docker file to tell GitHub what container to spin up.  The shell script that gets run by the container when it spins up, and this action file which defines the action.